Move pid and history files to project folder
To permit concurrent executions of reportupdater for different
projects, this change moves the pid file and the history file
(the two files that reportupdater uses to coordinate execution)
to the project folder (mobile, edit, etc.), instead of the
root folder. Thus, instances of the reportupdater for different
projects will run in parallel without collision.

Bug: T103385
Change-Id: I989618e3e5ff5e700b564d7191bc5f6a86dd05e6